models: Remove unused function can_access_all_realm_members.
This commit removes can_access_all_realm_members function as
it is not used anywhere in code other than tests.

This function was originally added in a003b5a74bbc and was
only used in digest.py other than the tests, but its use
in diget.py was removed in 634010088a74 and the function
itself was not removed from models.py.