hotspots: Remove useless and malformed alt text.
Since it was introduced in 12c35273e77140086e25677f488a9e6cf1c35b24,
we've had alt text for the hotspots illustration with an invalid
translation tag.

Fix this by removing the alt text, since the image has no functional
effect and so the right answer should it fail to load or the user is
visually impaired is to ignore it.